// src/modules/cache-manager/commons/cache-manager.enums.ts
export enum PREFIXES {
  LOGIN_ATTEMPT = 'LOGIN:',
  OTP_ATTEMPT = 'OTP_ATTEMPT:',
  OTP = 'OTP:',
  BLOCK_COUNT = 'BLOCK_COUNT:',
  BLOCK = 'BLOCK:',
  NOTIFICATION_COUNT = 'NOTIFICATION_COUNT:',
  ADMIN_INVITE_TOKEN = 'ADMIN_INVITE_TOKEN:',
  ADMIN_2FA_TOKEN = 'ADMIN_2FA_TOKEN:',
  ADMIN_FORGOT_PASSWORD = 'ADMIN_FORGOT_PASSWORD:',
}

export enum EXPIRES {
  LOGIN_ATTEMPT = 0,
  OTP_ATTEMPT = 0,
  OTP = 180,
  BLOCK_COUNT = 0,
  BLOCK_SIGNUP = 1800,
  BLOCK_LOGIN = 86400,
  FORGOT_OTP = 86400,
  ADMIN_INVITE_TOKEN = 86400,
  ADMIN_2FA_TOKEN = 1800,
  ADMIN_FORGOT_PASSWORD = 1800,
}